+++ front-end/src/views/admin/CategoryView.vue
@@ -78,75 +78,68 @@ const clearSearch = async () => {
 </script>
 
 <template>
- 
-   
 
-      <div class="flex justify-between items-center mb-6">
-        <h1 class="text-3xl font-bold text-white">Category Management</h1>
 
-        <div class="flex items-center gap-3">
-          <input
-            type="text"
-            placeholder="Search category..."
-            v-model="search"
-            class="border px-3 py-2 rounded w-64"
-          />
 
-          <button class="bg-white/10 hover:bg-white/20 text-white p-2 rounded-md">
-            <span class="pi pi-search"></span>
-          </button>
+  <div class="flex justify-between items-center mb-6">
+    <h1 class="text-3xl font-bold text-white">Category Management</h1>
 
-          <button class="bg-white/10 hover:bg-white/20 text-white p-2 rounded-md" @click="clearSearch">
-            <span class="pi pi-times"></span>
-          </button>
+    <div class="flex items-center gap-3">
+      <input type="text" placeholder="Search category..." v-model="search" class="border px-3 py-2 rounded w-64" />
 
-          <button
-            class="flex items-center gap-2 bg-[#637aad] hover:bg-[#4a68a8] text-white px-4 py-2 rounded-md"
-            @click="openCreateDialog"
-          >
-            <span class="pi pi-plus"></span>
-            Create
-          </button>
-        </div>
-      </div>
+      <button class="bg-white/10 hover:bg-white/20 text-white p-2 rounded-md">
+        <span class="pi pi-search"></span>
+      </button>
 
-      <!-- Table -->
-      <div class="bg-white rounded-lg overflow-hidden">
-        <table class="w-full text-left text-black">
-          <thead class="bg-[#383838] text-gray-300 text-sm">
-            <tr>
-              <th class="px-6 py-3">Category Name</th>
-              <th class="px-6 py-3 text-center">Action</th>
-            </tr>
-          </thead>
+      <button class="bg-white/10 hover:bg-white/20 text-white p-2 rounded-md" @click="clearSearch">
+        <span class="pi pi-times"></span>
+      </button>
 
-          <tbody class="divide-y">
+      <button class="flex items-center gap-2 bg-[#637aad] hover:bg-[#4a68a8] text-white px-4 py-2 rounded-md"
+        @click="openCreateDialog">
+        <span class="pi pi-plus"></span>
+        Create
+      </button>
+    </div>
+  </div>
 
-            <tr v-if="categoryStore.categories.length === 0">
-              <td colspan="2" class="text-center py-6 text-gray-500">
-                No data found
-              </td>
-            </tr>
+  <!-- Table -->
+  <div class="bg-white rounded-lg overflow-hidden">
+    <table class="w-full text-left text-black">
+      <thead class="bg-[#383838] text-gray-300 text-sm">
+        <tr>
+          <th class="px-6 py-3">หมวดหมู่ </th>
+          <th class="px-6 py-3 text-center">จัดการ</th>
+        </tr>
+      </thead>
 
-            <tr v-else v-for="category in categoryStore.categories" :key="category.category_id">
-              <td class="px-6 py-2">{{ category.name }}</td>
+      <tbody class="divide-y">
 
-              <td class="px-6 py-3 flex justify-center space-x-2">
-                <button @click="openEdit(category)">
-                  <span class="pi pi-pencil"></span>
-                </button>
+        <tr v-if="categoryStore.categories.length === 0">
+          <td colspan="2" class="text-center py-6 text-gray-500">
+            No data found
+          </td>
+        </tr>
 
-                <button @click="openDelete(category)">
-                  <span class="pi pi-trash"></span>
-                </button>
-              </td>
-            </tr>
+        <tr v-else v-for="category in categoryStore.categories" :key="category.category_id">
+          <td class="px-6 py-2">{{ category.name }}</td>
 
-          </tbody>
-        </table>
-      </div>
-   
- 
+          <td class="px-6 py-3 flex justify-center space-x-2">
+            <button @click="openEdit(category)">
+              <span class="pi pi-pencil"></span>
+            </button>
+
+            <button @click="openDelete(category)">
+              <span class="pi pi-trash"></span>
+            </button>
+          </td>
+        </tr>
+
+      </tbody>
+    </table>
+  </div>
+
+
   <!-- Dialog -->
   <div v-if="showDialog" class="overlay">
     <div class="dialog">
@@ -156,11 +149,8 @@ const clearSearch = async () => {
 
       <div class="mb-3">
         <label class="block mb-1">Category Name</label>
-        <input
-          v-model="categoryStore.editedCategory.name"
-          type="text"
-          class="border w-full px-3 py-2 rounded bg-gray-50"
-        />
+        <input v-model="categoryStore.editedCategory.name" type="text"
+          class="border w-full px-3 py-2 rounded bg-gray-50" />
       </div>
 
       <div class="flex justify-center gap-4">
@@ -175,21 +165,11 @@ const clearSearch = async () => {
     </div>
   </div>
 
-  <ConfirmComponent
-    :show="showConfirm"
-    type="save"
-    message="บันทึกข้อมูลนี้หรือไม่"
-    @confirm="saveCategory"
-    @cancel="showConfirm = false"
-  />
+  <ConfirmComponent :show="showConfirm" type="save" message="บันทึกข้อมูลนี้หรือไม่" @confirm="saveCategory"
+    @cancel="showConfirm = false" />
 
-  <ConfirmComponent
-    :show="deleteConfirm"
-    type="delete"
-    message="ต้องการลบข้อมูลนี้หรือไม่"
-    @confirm="removeItem(categoryStore.editedCategory)"
-    @cancel="closeDialogDelete"
-  />
+  <ConfirmComponent :show="deleteConfirm" type="delete" message="ต้องการลบข้อมูลนี้หรือไม่"
+    @confirm="removeItem(categoryStore.editedCategory)" @cancel="closeDialogDelete" />
 
   <LoadingComponent v-model="loadingStore.loading" />
 </template>
