+++ front-end/src/views/admin/UserView.vue
@@ -2,11 +2,15 @@
 import { useUserStore } from '@/stores/user'
 import { onMounted, ref } from 'vue'
 import ConfirmComponent from '@/components/dialogs/ConfirmComponent.vue'
-
 import type { User } from '@/types/User'
 import { useAuthStore } from '@/stores/auth'
 import { useLoadingStore } from '@/stores/loading'
 import LoadingComponent from '@/components/LoadingComponent.vue'
+import { computed } from 'vue'
+import { useBranchStore } from '@/stores/branch'
+
+
+
 const loadingStore = useLoadingStore()
 const userStore = useUserStore()
 const showDialog = ref(false)
@@ -14,13 +18,17 @@ const search = ref('')
 const showConfirm = ref(false)
 const deleteConfirm = ref(false)
 const authStore = useAuthStore()
+const branchStore = useBranchStore()
 
 onMounted(async () => {
   await userStore.getUsers()
   await authStore.getCurrentUser()
+  await branchStore.getBranches()
 })
 
-import { computed } from 'vue'
+
+
+
 
 const formUser = computed(() => {
   return mode.value === 'create' ? userStore.createUser : userStore.editedUser
@@ -37,27 +45,19 @@ const openEdit = (user: User) => {
   }
 
   if (authStore.user?.role === 'manager') {
-    userStore.createUser.role = 'employee'
+    userStore.createUser.role = 'staff'
   }
   showDialog.value = true
 }
 
-// const saveUser = () => {
-//   if (mode.value === 'create') {
-//     userStore.addUser(userStore.editedUser)
-//   } else {
-//     userStore.updateUser(userStore.editedUser)
-//   }
-//   showDialog.value = false
-//   userStore.clearBrance()
-//   userStore.getUseres()
-// }
+
 
 const saveUser = async () => {
   console.log(mode.value)
   if (mode.value === 'create') {
     console.log(userStore.createUser)
     await userStore.addUser(userStore.createUser)
+    console.log(userStore.createUser)
     userStore.clearCreateUser()
   } else if (mode.value === 'edit' && userStore.editedUser) {
     await userStore.updateUser(userStore.editedUser)
@@ -82,7 +82,7 @@ const openCreateDialog = () => {
   }
 
   if (authStore.user?.role === 'manager') {
-    userStore.createUser.role = 'employee'
+    userStore.createUser.role = 'staff'
   }
 
   showDialog.value = true
@@ -107,10 +107,7 @@ const searchUser = async () => {
   await userStore.getUsers()
 }
 
-// const openDelete = async (item: User) => {
-//   deleteConfirm.value = true
-//   userStore.editedUser = { ...item }
-// }
+
 
 const closeDialogDelete = async () => {
   deleteConfirm.value = false
@@ -126,95 +123,84 @@ const clearSearch = async () => {
 </script>
 
 <template>
- 
 
-      <div class="flex justify-between items-center mb-6">
-        <h1 class="text-3xl font-bold text-white">User Management</h1>
 
-        <div class="flex items-center gap-3">
-          <input
-            type="text"
-            placeholder="Search User..."
-            v-model="search"
-            class="border px-3 py-2 rounded w-64"
-          />
+  <div class="flex justify-between items-center mb-6">
+    <h1 class="text-3xl font-bold text-white">User Management</h1>
 
-          <button
-            class="bg-white/10 hover:bg-white/20 text-white p-2 rounded-md flex items-center justify-center"
-            @click="searchUser()"
-          >
-            <span class="pi pi-search text-lg"></span>
-          </button>
-          <button
-            class="bg-white/10 hover:bg-white/20 text-white p-2 rounded-md flex items-center justify-center"
-            @click="clearSearch()"
-          >
-            <span class="pi pi-times text-lg"></span>
-          </button>
-          <button
-            class="flex items-center gap-2 bg-[#637aad] hover:bg-[#4a68a8]  text-white px-4 py-2 rounded-md  transition"
-            @click="openCreateDialog()"
-          >
-            <span class="pi pi-plus text-lg"></span>
-            <span>Create</span>
-          </button>
-        </div>
-      </div>
+    <div class="flex items-center gap-3">
+      <input type="text" placeholder="Search User..." v-model="search" class="border px-3 py-2 rounded w-64" />
 
-    
-      <div class="bg-white rounded-lg overflow-hidden">
-        <table class="w-full text-left text-black">
-          <thead class="bg-[#383838] text-gray-300 text-sm">
-            <tr>
-              <th class="px-6 py-3">User Name</th>
-              <th class="px-6 py-3">Email</th>
-              <th class="px-6 py-3">Role</th>
-              <th class="px-6 py-3">Enabled</th>
-              <th class="px-6 py-3">Branch</th>
-              <th class="px-6 py-3 text-center">Action</th>
-            </tr>
-          </thead>
+      <button class="bg-white/10 hover:bg-white/20 text-white p-2 rounded-md flex items-center justify-center"
+        @click="searchUser()">
+        <span class="pi pi-search text-lg"></span>
+      </button>
+      <button class="bg-white/10 hover:bg-white/20 text-white p-2 rounded-md flex items-center justify-center"
+        @click="clearSearch()">
+        <span class="pi pi-times text-lg"></span>
+      </button>
+      <button
+        class="flex items-center gap-2 bg-[#637aad] hover:bg-[#4a68a8]  text-white px-4 py-2 rounded-md  transition"
+        @click="openCreateDialog()">
+        <span class="pi pi-plus text-lg"></span>
+        <span>Create</span>
+      </button>
+    </div>
+  </div>
 
-          <tbody class="divide-y">
-            <tr v-if="userStore.users.length === 0">
-              <td colspan="4" class="text-center py-6 text-gray-500">ไม่พบข้อมูลที่ค้นหา</td>
-            </tr>
 
-            <tr v-else v-for="User in userStore.users" :key="User.user_id">
-              <td class="px-6 py-1">{{ User.name }}</td>
-              <td class="px-6 py-1">{{ User.email }}</td>
-              <td class="px-6 py-1">{{ User.role }}</td>
-              <td class="px-6 py-1">{{ User.enabled }}</td>
-              <td class="px-6 py-1">{{ User.email }}</td>
+  <div class="bg-white rounded-lg overflow-hidden">
+    <table class="w-full text-left text-black">
+      <thead class="bg-[#383838] text-gray-300 text-sm">
+        <tr>
+          <th class="px-6 py-3">ชื่อผู้ใช้</th>
+          <th class="px-6 py-3">อีเมล</th>
+          <th class="px-6 py-3">บทบาท</th>
+          <th class="px-6 py-3">สถานะการใช้งาน</th>
+          <th class="px-6 py-3">สาขา</th>
+          <th class="px-6 py-3 text-center">จัดการ</th>
+        </tr>
+      </thead>
 
-           
+      <tbody class="divide-y">
+        <tr v-if="userStore.users.length === 0">
+          <td colspan="4" class="text-center py-6 text-gray-500">ไม่พบข้อมูลที่ค้นหา</td>
+        </tr>
 
-              <td class="px-6 py-3 flex justify-center space-x-2">
-                <button class="edit-btn" @click="openEdit(User)">
-                  <span class="pi pi-pencil"></span>
-                </button>
-              </td>
-            </tr>
-          </tbody>
-        </table>
+        <tr v-else v-for="User in userStore.users" :key="User.user_id">
+          <td class="px-6 py-1">{{ User.name }}</td>
+          <td class="px-6 py-1">{{ User.email }}</td>
+          <td class="px-6 py-1">{{ User.role }}</td>
+          <td class="px-6 py-1">{{ User.enabled }}</td>
+          <td class="px-6 py-1">{{ User.email }}</td>
 
-        <!-- Pagination -->
-        <div class="flex justify-end items-center gap-4 py-4 border-t mr-3">
-          <button class="px-3 py-1 border rounded hover:bg-gray-100" @click="prevPage()">
-            <span class="pi pi-chevron-left text-sm"></span> Prev
-          </button>
 
-          <span class="text-sm text-gray-600">
-            {{ userStore.page }} of {{ userStore.lastPage }}</span
-          >
 
-          <button class="px-3 py-1 border rounded hover:bg-gray-100" @click="nextPage()">
-            Next <span class="pi pi-chevron-right text-sm"></span>
-          </button>
-        </div>
-      </div>
+          <td class="px-6 py-3 flex justify-center space-x-2">
+            <button class="edit-btn" @click="openEdit(User)">
+              <span class="pi pi-pencil"></span>
+            </button>
+          </td>
+        </tr>
+      </tbody>
+    </table>
 
- 
+    <!-- Pagination -->
+    <div class="flex justify-end items-center gap-4 py-4 border-t mr-3">
+      <button class="px-3 py-1 border rounded hover:bg-gray-100" @click="prevPage()">
+        <span class="pi pi-chevron-left text-sm"></span> Prev
+      </button>
+
+      <span class="text-sm text-gray-600">
+        {{ userStore.page }} of {{ userStore.lastPage }}</span>
+
+      <button class="px-3 py-1 border rounded hover:bg-gray-100" @click="nextPage()">
+        Next <span class="pi pi-chevron-right text-sm"></span>
+      </button>
+    </div>
+  </div>
+
+
 
   <!-- Dialog -->
   <div v-if="showDialog && formUser" class="overlay">
@@ -226,53 +212,43 @@ const clearSearch = async () => {
       <!-- Name -->
       <div class="mb-3">
         <label>User Name</label>
-        <input
-          v-model="formUser.name"
-          type="text"
-          placeholder="Enter user name"
-          class="border w-full px-3 py-2 rounded bg-gray-50"
-        />
+        <input v-model="userStore.createUser.name" type="text" placeholder="Enter user name"
+          class="border w-full px-3 py-2 rounded bg-gray-50" />
       </div>
 
-      <!-- Email (create only) -->
+      <!-- Email  -->
       <div v-if="mode === 'create'" class="mb-3">
         <label>Email</label>
-        <input
-          v-model="userStore.createUser.email"
-          type="email"
-          placeholder="Enter email"
-          class="border w-full px-3 py-2 rounded bg-gray-50"
-        />
+        <input v-model="userStore.createUser.email" type="email" placeholder="Enter email"
+          class="border w-full px-3 py-2 rounded bg-gray-50" />
       </div>
 
       <div v-if="mode === 'edit' && userStore.editedUser" class="mb-3">
         <label>Email</label>
-        <input
-          v-model="userStore.editedUser.email"
-          type="email"
-          placeholder="Enter email"
-          class="border w-full px-3 py-2 rounded bg-gray-50"
-        />
+        <input v-model="userStore.editedUser.email" type="email" placeholder="Enter email"
+          class="border w-full px-3 py-2 rounded bg-gray-50" />
       </div>
       <div v-if="mode === 'create'" class="mb-3">
         <label>Password</label>
-        <input
-          v-model="userStore.createUser.password"
-          type="password"
-          placeholder="Enter password"
-          class="border w-full px-3 py-2 rounded bg-gray-50"
-        />
+        <input v-model="userStore.createUser.password" type="password" placeholder="Enter password"
+          class="border w-full px-3 py-2 rounded bg-gray-50" />
       </div>
 
       <div class="mb-3">
         <label>Role</label>
-        <input
-          v-model="formUser.role"
-          class="border w-full px-3 py-2 rounded bg-gray-200"
-          disabled
-        />
+        <input v-model="formUser.role" class="border w-full px-3 py-2 rounded bg-gray-200" disabled />
       </div>
+      <div class="mb-3">
+        <label>Branch</label>
 
+        <select v-model="userStore.createUser.branch_id" class="border w-full px-3 py-2 rounded bg-gray-50">
+          <option value="" disabled>Select branch</option>
+
+          <option v-for="b in branchStore.branches" :key="b.branch_id" :value="b.branch_id">
+            {{ b.branch_name }}
+          </option>
+        </select>
+      </div>
       <div v-if="mode === 'edit' && userStore.editedUser" class="mb-3">
         <label>Enabled</label>
         <select v-model="userStore.editedUser.enabled" class="border w-full px-3 py-2 rounded">
@@ -293,20 +269,12 @@ const clearSearch = async () => {
     </div>
   </div>
 
-  <ConfirmComponent
-    :show="showConfirm"
-    type="save"
-    message="คุณต้องการที่จะบันทึกข้อมูลนี้"
-    @confirm="saveUser()"
-    @cancel="showConfirm = false"
-  />
+  <ConfirmComponent :show="showConfirm" type="save" message="คุณต้องการที่จะบันทึกข้อมูลนี้" @confirm="saveUser()"
+    @cancel="showConfirm = false" />
 
-  <ConfirmComponent
-    :show="deleteConfirm"
-    type="delete"
-    message="คุณต้องการที่จะลบข้อมูลนี้ใช่หรือไม่"
-    @cancel="closeDialogDelete()"
-  /> <LoadingComponent v-model="loadingStore.loading" />
+  <ConfirmComponent :show="deleteConfirm" type="delete" message="คุณต้องการที่จะลบข้อมูลนี้ใช่หรือไม่"
+    @cancel="closeDialogDelete()" />
+  <LoadingComponent v-model="loadingStore.loading" />
 </template>
 
 <style scoped>
